Updated sv_SE.jsonc
Updated sv_SE.jsonc by correcting spelling and correcting my word choice after using the language for a few days in Cider so it makes sense in the context. 

This will probably be the last edit until more terms and such are added as the language should be good for now.