Improve Smooth Lyric Animation (#1096)
* app-mainMenu can be triggered by right click

Currently you can't open the little Cider popup menu without first uncollapsing the sidebar. This is extremely an extremely unepic sussy amogus L gamer moment. This change allows you to right click to bypass uncollapsing the sidebar and get straight to the hot app-mainMenu action. I think Cider speedrunners will greatly appreciate this quality of life addition. Please consider accepting this PR (short for pull request) to improve the UX (short for user experience).

* Improve Smooth Lyric Animation

PR #1094 introduced a few issues:
A) The animations would be paused when Cider was out of focus, making the lyrics remain grayed out
B) The animation was too slow, making it harder to sing along with karaoke mode on faster-paced songs
C) The easing method was inconsistent with the rest of the app and made the color look a little janky

This PR fixes issue A by changing the animation method from an animation to a transition, fixes B by increasing the speed (which themes can easily adjust if preferred), and fixes C by using the existing var(--appleEase) as the easing method